queue/pyproject.toml
[project]
name = "queue"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "browser-use>=0.1.48",
    "langchain-groq>=0.3.2",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
]
